Webhook Dispatcher 0.1.0 Initial Commit
This is a webhook dispatcher for taking github org hooks and manually triggering
concourse pipelines.

This allows one github org hook to trigger all your pipelines instead of having
concourse poll your github org. The problem with polling is that it uses up all
of your github API request limit as the number of pipelines increases.